+++ CHANGES.rst
@@ -68,6 +68,8 @@ Release history
   or ``RectifiedLinear`` as the base type (these are equivalent to ``LIF`` and
   ``SpikingRectifiedLinear``, respectively).
   (`#296 <https://github.com/nengo/nengo-loihi/pull/296>`__)
+- Added support for ``padding='same'`` on ``nengo.Convolution`` transforms.
+  (`#297 <https://github.com/nengo/nengo-loihi/pull/297>`__)
 
 **Changed**